Lannister/api_src/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
python-dotenv
pytest==7.1.2
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1